requirements.txt
python-telegram-bot>=20.0
jieba>=0.42.1
aiohttp>=3.8.0
dashscope>=1.14.0
python-dotenv>=1.0.0
